fix error

+++ ssc/Workspaces/workspaces.py
@@ -405,6 +405,8 @@ def decrypt_file(workspace_name, filename, audio_key):
     altered = str(split[0] + time_stamp + 'decrypted' + split[2])
     key_string = bytes(audio_key, 'utf-8')
     encoded_key = base64.b64encode(key_string)
+    generic_error=None
+    to_send=None
     try:
 
         s3.Bucket(workspace_name).download_file(filename, altered_filename)
@@ -429,15 +431,20 @@ def decrypt_file(workspace_name, filename, audio_key):
 
     except (Exception, psycopg2.Error) as error:
         print('Error while connecting to PostgresQL', error)
+        generic_error = 1
+    except:
+        generic_error=1
     finally:
         if (connection):
             cursor.close()
             connection.close()
             print("PostgresSQL connection is closed")
-
-        to_send = send_file(altered)
-        os.remove(altered)
-        return to_send
+        if(generic_error==1):
+            return {"incorrect_key": True}
+        else:
+            to_send = send_file(altered)
+            os.remove(altered)
+            return to_send
 
 
 def fetch_workspace_users(name):
